requirements.txt
numpy
pandas
pyarrow
rdkit-pypi
scikit-learn
